+++ src/components/search/SearchClass.tsx
@@ -4,7 +4,6 @@ import ClassPropertyDisplay from '@/components/plan/ClassPropertyDisplay';
 import { openInfo } from '@/components/plan/CourseInfo';
 import { Color } from '@/types/BaseTypes';
 import { Course, CourseDragItem } from '@/types/PlanTypes';
-import { convertDisciplines, convertDistros } from '@/utility/Utility';
 import { BookmarkIcon, ChevronRightIcon } from '@heroicons/react/24/outline';
 import { BookmarkIcon as BookmarkIconSolid } from '@heroicons/react/24/solid';
 import { motion } from 'framer-motion';
@@ -37,8 +36,6 @@ function SearchClass({ course, color, select, selected }: SearchClassProps) {
   );
 
   const recentOfferings = getOfferings(course).slice(0, 8);
-  const distroStrings = convertDistros(course.distros);
-  const disciplinesStrings = convertDisciplines(course.disciplines);
   const units = parseFloat(course.units);
   const isFavorite = bookmarks?.noCredit.has(course);
   return (
@@ -60,12 +57,6 @@ function SearchClass({ course, color, select, selected }: SearchClassProps) {
         {course.id}
       </p>
       <p className="text-sm text-black dark:text-gray-50">{course.name}</p>
-      <p className="mt-4 text-xs text-gray-700 dark:text-gray-300">
-        {course.description}
-      </p>
-      {course.prereqs && (
-        <ClassPropertyDisplay title="PREREQUISITES" value={course.prereqs} />
-      )}
       <ClassPropertyDisplay
         title="RECENT OFFERINGS"
         value={
@@ -74,18 +65,6 @@ function SearchClass({ course, color, select, selected }: SearchClassProps) {
             : 'Not offered recently'
         }
       />
-      {disciplinesStrings.length > 0 && (
-        <ClassPropertyDisplay
-          title="FOUNDATIONAL DISCIPLINES"
-          value={disciplinesStrings}
-        />
-      )}
-      {distroStrings.length > 0 && (
-        <ClassPropertyDisplay
-          title="DISTRIBUTION AREAS"
-          value={distroStrings}
-        />
-      )}
       <div className="mt-1">
         <p className="text-right text-xs font-light text-gray-500 dark:text-gray-400">
           <span className="font-medium">{units}</span>{' '}
@@ -99,6 +78,16 @@ function SearchClass({ course, color, select, selected }: SearchClassProps) {
           animate={{ opacity: 1 }}
           transition={{ delay: 0.25, duration: 0.2 }}
         >
+          <button
+            className="mx-auto my-2 flex items-center rounded-sm px-1 py-0.5 text-xs font-bold text-gray-500 hover:bg-black/5 active:bg-black/10 dark:text-gray-400 dark:hover:bg-white/5 dark:active:bg-white/10"
+            onClick={(e) => {
+              e.stopPropagation();
+              openInfo(course, app, true);
+            }}
+          >
+            <span>VIEW MORE INFO</span>
+            <ChevronRightIcon className="h-4 w-4 stroke-2" />
+          </button>
           <AddButtons
             action={(year, quarter) => {
               planModification.addCourse(course, {
@@ -140,16 +129,6 @@ function SearchClass({ course, color, select, selected }: SearchClassProps) {
                 : 'Add for credit'}
             </button>
           </div>
-          <button
-            className="mx-auto my-2 flex items-center text-xs font-bold text-gray-400 hover:text-gray-500 active:text-gray-600 dark:text-gray-500 dark:hover:text-gray-400 dark:active:text-gray-300"
-            onClick={(e) => {
-              e.stopPropagation();
-              openInfo(course, app, true);
-            }}
-          >
-            <span>VIEW MORE INFO</span>
-            <ChevronRightIcon className="h-4 w-4 stroke-2" />
-          </button>
         </motion.div>
       )}